refactor(server): extract breakpoint tools into src/server/handlers/breakpoint-tools.ts
set_breakpoint, list_breakpoints, remove_breakpoint and clear_breakpoints
move off the CallTool switch; bodies verbatim with `this.` -> `ctx.` and
`result = X; break;` -> `return X;`.

set_breakpoint splits at its existing `if (isFunctionBp)` seam: the shared
entry guard and addressing-mode gating stay in setBreakpointTool, and the
two arms become setFunctionBreakpointBranch / setLineBreakpointBranch. The
branch functions take args narrowed to `sessionId: string` (a call-site cast,
since TypeScript narrows the property after the guard but not the object), so
neither body needed a single `!` added. The fn-branch's `break` after its
try/catch is the return of each arm; the line branch keeps its post-assignment
tool:set_breakpoint:result log by holding the result in a const and returning
it after the log, inside the same try. remove_breakpoint's three early
`break`s inside the try become returns, which exit the handler the same way.
